inclusao de state do objeto da aplicação
